Modified first changes to rmdir

+++ tbfs.py
@@ -125,9 +125,11 @@ class MyFS(fuse.Fuse):
         return -errno.EEXIST    # if the path is already in path
 
     def rmdir(self, path):
-        full_path = self._full_path(path)
-        return os.rmdir(full_path)
+        if path in self.hash_dict:
+            return os.rmdir(full_path)
 
+        return -errno.EEXIST    # if the path does not exist
+        
     def chmod(self, path, mode):
         print "*****CHMOD: ",path
         if path in self.hash_dict:
